# swagger.yml
---
openapi: 3.0.3
info: 
  version: 2.22.0
  title:
    Multimedia Projects
  description: |-
    This is a simpe website for IVKHK Mutlimedia projects. This site implements search and viewing of projects that were created by students of multimedia group

paths: 
  /:
    get:
      tags:
        - Home Page
      description: Index Page
      responses:
        '200':
          description: OK. Response home.html
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: A page not found.
        '5XX':
          description: Unexpected error.


  /home:
  
    get:
      tags:
        - Home Page
      description: Index Page
      responses:
        '200':
          description: OK. Response home.html
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: A page not found.
        '5XX':
          description: Unexpected error.

  /admin/home:
    
    get:
      tags:
        - Home Page
      description: Главная страница админа
      responses:
        '200':
          description: OK. Response AdminHome.html
        '400':
          description: Bad request. User ID must be an integer and larger than 0.
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: A page not found.
        '5XX':
          description: Unexpected error.
        
  /result:
    
    get:
      tags:
        - Result Page
      summary: Result page
      description: A page request 2 parametrs for searching and response 'results' - array of students in json
      parameters:
        - schema:
            type: string
          in: query
          name: message
          description: 'A message for searchig'
        - schema:
            type: string
          in: query
          name: key
          description: 'A key for searching'
      responses:
        '200':
          description: OK. Response a list of students
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    description: Student name
                  group:
                    type: string
                    description: Student group
                  project:
                    type: string
                    description: Student project
                  typeproject:
                    type: string
                    description: Project type
                  year:
                    type: string
                    description: Project year
                  uuid:
                    type: string
                    description: Project uuid
        '400':
          description: Bad request. A fields are filled in incorrectly.
        '404':
          description: A page not found.
        '5XX':
          description: Unexpected error.

  /project:
    
    get:
      tags:
        - Project
      summary: Project Page
      description: An admin page with info about project. A page request 2 parametrs for searching and response 'results' - array of project info in json
      parameters:
        - schema:
            type: string
          in: query
          name: project
          description: 'A project name'
        - schema:
            type: string
          in: query
          name: uuid
          description: 'A uuid of project'
      responses:
        '200':
          description: OK. Response a list of students
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    description: Student name
                  group:
                    type: string
                    description: Student group
                  project:
                    type: string
                    description: Project name
                  typeproject:
                    type: string
                    description: Project type
                  year:
                    type: string
                    description: Project year
                  uuid:
                    type: string
                    description: Project uuid
                  project_file:
                    type: string
                    description: A project file
                  app_file_primary:
                    type: string
                    description: First file of project
                  app_files:
                    type: array
                    description: Rest file of project
        '400':
          description: Bad request. Еhe fields are filled in incorrectly.
        '404':
          description: A page not found.
        '5XX':
          description: Unexpected error.


  '/admin/project':
    
    get:
      tags:
        - Project
      summary: Project Page
      description: A page with info about project. A page request 2 parametrs for searching and response 'results' - array of project info in json
      parameters:
        - schema:
            type: string
          in: query
          name: project
          description: 'A project name'
        - schema:
            type: string
          in: query
          name: uuid
          description: 'A uuid of project'
      responses:
        '200':
          description: OK. Response a list of students
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    description: Student name
                  group:
                    type: string
                    description: Student group
                  project:
                    type: string
                    description: Project name
                  typeproject:
                    type: string
                    description: Project type
                  year:
                    type: string
                    description: Project year
                  uuid:
                    type: string
                    description: Project uuid
                  project_file:
                    type: string
                    description: A project file
                  app_file_primary:
                    type: string
                    description: First file of project
                  app_files:
                    type: array
                    description: Rest file of project
        '400':
          description: Bad request. Еhe fields are filled in incorrectly.
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: A page not found.
        '5XX':
          description: Unexpected error.

  '/admin/add':
    
    get:
      tags:
        - Add project
      description: Страница добавления проекта
      responses:
        '200':
          description: Successful operation

  '/admin/edit':
    
    get:
      tags:
        - Edit project
      description: Страница редактирования проекта
      responses:
        '200':
          description: Successful operation

  '/admin/delete':
    
    get:
      tags:
        - Delete project
      description: Страница удаления проекта
      responses:
        '200':
          description: Successful operation